# speedup wf{3,4}_of_wf by factorizing raw matches to definitions