This greatly reduces term size (tree size 1988106 -> 1121651 for wf3_of_wf).

Timings before:
(Time Succeed Qed includes some time which isn't in Time Qed, see also https://github.com/coq/coq/pull/19426)

wf3
tactics 2.2s
qed 1.09s
succeed qed 1.25s

wf4
tactics 14.5s
qed 7.4s
succeed qed 8.5s

After:

wf3
tactics 1.4s
qed 0.6s
succeed qed 0.65s

wf4
tactics 8s
qed 3.6s
succeed qed 4s