@JasonGross 2024-07-29
Indeed, I'm guessing the docker image or the apt repositories were recently updated, let's see if https://github.com/mit-plv/rewriter/pull/158 fixes it